Some fixes to make the code complile
- The nurbs++/matrix library can now be used
- The nurbs++/nurbs library can now be used
- Some of the examples complile

There is a problem with the friend declarations in galloc.h and galloc2d.h.
The workaround is to make variables in barray.h and barray2d.h public!
This is no longterm solution but makes the library usable again.

The correct functionality of the library has not been tested yet!
There are still some unresolved variables in the libraries.